creature/behavior: introduce basic creature behavior module
