+++ src/trx/game/creature/common.c
@@ -26,8 +26,6 @@
     (g_TRVersion == 1 ? FRONT_ARC : 0x3000) // = 16384 (TR1), 12288 (TR2)
 #define M_MAX_X_ROT (20 * DEG_1) // = 3640
 
-static bool m_AlliesHostile = false;
-
 static bool M_TestSwitchOrKill(
     const int16_t item_num, const OBJECT_ID target_id)
 {
@@ -1035,16 +1033,6 @@ int32_t Creature_Vault(
     return vault;
 }
 
-bool Creature_AreAlliesHostile(void)
-{
-    return m_AlliesHostile;
-}
-
-void Creature_SetAlliesHostile(bool enable)
-{
-    m_AlliesHostile = enable;
-}
-
 bool Creature_IsAlive(const ITEM *const item)
 {
     const OBJECT *const obj = Object_Get(item->object_id);
@@ -1054,22 +1042,6 @@ bool Creature_IsAlive(const ITEM *const item)
 
     return (item->hit_points > 0)
         || (item->hit_points == DONT_TARGET && item->active);
-}
-
-bool Creature_IsHostile(const ITEM *const item)
-{
-    return Object_IsType(item->object_id, g_EnemyObjects)
-        || (Creature_AreAlliesHostile() && Creature_IsAlly(item));
-}
-
-bool Creature_IsAlly(const ITEM *const item)
-{
-    return Object_IsType(item->object_id, g_AllyObjects);
-}
-
-bool Creature_IsAllyTargetingEnemy(const ITEM *const item)
-{
-    return Object_IsType(item->object_id, g_AllyTargetingEnemies);
 }
 
 bool Creature_IsTargetable(const ITEM *const item)
